+++ Arbiter.App/ViewModels/Proxy/ProxyViewModel.DebugMessage.cs
@@ -1,4 +1,7 @@
-﻿using Arbiter.App.Models;
+﻿using System.Collections.Generic;
+using System.Linq;
+using System.Text.RegularExpressions;
+using Arbiter.App.Models;
 using Arbiter.Net;
 using Arbiter.Net.Filters;
 using Arbiter.Net.Proxy;
@@ -9,9 +12,12 @@ namespace Arbiter.App.ViewModels.Proxy;
 
 public partial class ProxyViewModel
 {
+    private const string DebugWorldMessageFilterName = "Debug_WorldMessageFilter";
     private const string DebugEmptyWorldMessageFilterName = "Debug_EmptyWorldMessageFilter";
 
-    private void AddDebugMessageFilters(DebugSettings settings)
+    private readonly Dictionary<string, Regex> _regexCache = new();
+    
+    private void AddDebugMessageFilters(DebugSettings settings, IReadOnlyList<MessageFilter> filters)
     {
         if (settings.IgnoreEmptyMessages)
         {
@@ -22,11 +28,54 @@ public partial class ProxyViewModel
                     Priority = int.MaxValue
                 });
         }
+        
+        if (filters.Count > 0)
+        {
+            _proxyServer.AddFilter(ServerCommand.WorldMessage,
+                new NetworkPacketFilter(HandleWorldMessage, filters.ToList())
+                {
+                    Name = DebugWorldMessageFilterName,
+                    Priority = int.MaxValue - 1
+                });
+        }
     }
 
     private void RemoveDebugMessageFilters()
     {
         _proxyServer.RemoveFilter(ServerCommand.WorldMessage, DebugEmptyWorldMessageFilterName);
+        _proxyServer.RemoveFilter(ServerCommand.WorldMessage, DebugWorldMessageFilterName);
+        
+        _regexCache.Clear();
+    }
+
+    private NetworkPacket? HandleWorldMessage(ProxyConnection connection, NetworkPacket packet, object? parameter)
+    {
+        // Ensure the packet is the correct type and we have settings as a parameter
+        if (packet is not ServerPacket serverPacket || parameter is not IReadOnlyList<MessageFilter> filters)
+        {
+            return packet;
+        }
+
+        if (!_serverMessageFactory.TryCreate<ServerWorldMessageMessage>(serverPacket, out var message))
+        {
+            return packet;
+        }
+
+        foreach (var filter in filters)
+        {
+            if (!_regexCache.TryGetValue(filter.Pattern, out var regex))
+            {
+                regex = new Regex(filter.Pattern, RegexOptions.IgnoreCase);
+                _regexCache[filter.Pattern] = regex;
+            }
+
+            if (regex.IsMatch(message.Message))
+            {
+                return null;
+            }
+        }
+
+        return packet;
     }
 
     private NetworkPacket? HandleEmptyWorldMessageMessage(ProxyConnection connection, NetworkPacket packet,